requirements/dev.txt
-r base.txt
ipython==8.25.0
psycopg==3.1.19
psycopg-binary==3.1.19
django-multiselectfield==0.1.13